requirements.txt
python-telegram-bot==20.7
gspread
oauth2client
python-dotenv
matplotlib

